backend-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
mysql-connector-python==8.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.27.0
joblib==1.3.2
transformers==4.46.1
torch==2.6.0
torchaudio==2.6.0
peft==0.7.1
accelerate==0.25.0
# chatterbox-tts>=0.1.6
openai-whisper>=20231117
soundfile>=0.12.1

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Piper TTS (Fast ONNX)
piper-tts>=1.2.0
onnxruntime>=1.16.0

